requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-amazon==9.3.0
cosmos-airflow-dbt==0.2.3
pandas==2.1.1
requests==2.31.0